+++ includes/modules/class-memberships-module.php
@@ -95,6 +95,13 @@ class Memberships_Module extends Module_Base {
 	}
 
 	/**
+	 * {@inheritDoc}
+	 */
+	public function get_required_modules(): array {
+		return [ 'woocommerce' ];
+	}
+
+	/**
 	 * Boot the module: register WC Memberships hooks.
 	 *
 	 * @return void