+++ includes/modules/class-wc-bookings-module.php
@@ -91,6 +91,13 @@ class WC_Bookings_Module extends Booking_Module_Base {
 	}
 
 	/**
+	 * {@inheritDoc}
+	 */
+	public function get_required_modules(): array {
+		return [ 'woocommerce' ];
+	}
+
+	/**
 	 * Boot the module: register WC Bookings hooks.
 	 *
 	 * @return void