+++ includes/modules/class-wc-points-rewards-module.php
@@ -89,6 +89,13 @@ class WC_Points_Rewards_Module extends Module_Base {
 	}
 
 	/**
+	 * {@inheritDoc}
+	 */
+	public function get_required_modules(): array {
+		return [ 'woocommerce' ];
+	}
+
+	/**
 	 * Boot the module: register WC Points & Rewards hooks.
 	 *
 	 * @return void