+++ includes/modules/class-wc-subscriptions-module.php
@@ -111,6 +111,13 @@ class WC_Subscriptions_Module extends Module_Base {
 	}
 
 	/**
+	 * {@inheritDoc}
+	 */
+	public function get_required_modules(): array {
+		return [ 'woocommerce' ];
+	}
+
+	/**
 	 * Boot the module: register WC Subscriptions hooks.
 	 *
 	 * @return void